more bugs fixed
add function to ensure the direction of the way is preserved when inserting new points (could sometimes be messed up when two ways looped around each other and had multiple intersections (ex, a V intersecting a square))
advance position along current way being intersected for intersections by number of intersections with the other way, not just by 1
always chose first two points assuming we'd be at an intersection, and the next point would be a non intersecting point, which was wrong. this is fixed.