fixed incorrect tracing when waysContainingPoint.Count only had viewPort as multiple entries
due to adding the .Equals method in the XYPair class, now, the first and
last point of the viewPort would not be considered separate Points, such
that waysContainingPoint would only have 1 result for them (ie the
viewPort) as before adding the .Equals method.
Now, they would have two results - the viewPort twice. This would break
the current logic, as we would not simply choose the viewPort to follow
(as when waysContainingPoint.Count == 1), but would instead continue to
follow the same current way, since none of the ways containg the point
wouldn't be the viewPort...