handle case of just backtracked into a point that touches but doesn't intersect viewport
in this case, we will have two possible polygons able to be formed, which intersect
at a single point. we can't remove this point from the intersections list, as
we still need it to form the second polygon. Removing it means second polygon
isn't able to be formed, and we get endless loop

an example happens at -14, 143