improve support for points touching but not transecting
now supports points where the next or previous are on the viewport edge

to enable this, had to modify PointInViewPortEdge logic. The prior PointInViewPortEdge
was changed to PointOnViewPortSegment.

Modified CleanOutsideSinglePointIntersections logic as well to support all of this
Had to use SafeEquals functions in GetDelta and PointOnViewPortSegment
in order to support all of this